add support content in comments, refactoring content component
